BACKEND/appointment_management/read_service/requirements.txt
fastapi==0.95.2
pydantic==1.10.2
mysql-connector-python==8.0.32
python-dotenv==0.21.0
uvicorn==0.22.0
pymysql
requests
PyJWT==2.6.0
